mer 19 lug 2017, 03.27.54, CEST

+++ modules/fs/fat/fat_read.c
@@ -38,7 +38,7 @@ int fat_read(inode_t* ino, void* buf, size_t size) {
 	int current_cluster = fat->entry_cluster;
 
 
-    int i, s, c;
+    long i, s, c;
     for(i = 0; i < (ino->position / (fat->sector_per_cluster * fat->bytes_per_sector)); i++) {
         current_cluster = fatutils_next_cluster(fat, current_cluster);
         switch(current_cluster) {
@@ -62,7 +62,7 @@ int fat_read(inode_t* ino, void* buf, size_t size) {
         char* p = (char*) buf; 
         fat->dev->position = (off64_t)((CLUSTER_TO_SECTOR(fat, current_cluster) * fat->bytes_per_sector) +
                                 ((int)ino->position % (fat->sector_per_cluster * fat->bytes_per_sector)));
-        
+
         if(i + c < s) {
             zero_if(vfs_read(fat->dev, &p[i], c) != c, EIO);
             i += c;
